[WFCORE-6645] Move HostReloadProxyTestCase to Domain Mode Test suite
