+++ testsuite/domain/src/test/java/org/jboss/as/test/integration/domain/suites/DomainTestSuite.java
@@ -23,6 +23,7 @@ import org.junit.runners.Suite;
         HostSuspendResumeTestCase.class,
         SyncModelOperationTestCase.class,
         ManagementReadXmlTestCase.class,
+        HostReloadProxyTestCase.class,
         AuditLogTestCase.class,
         IgnoredResourcesProfileCloneTestCase.class,
         CapabilityReloadRequiredUnitTestCase.class,
